wip; saving intermediate work, mostly notes for a plan
